refactor authentication routes
