+++ apps/personal-blog-backend/src/app/routes/index.ts
@@ -1,9 +1,11 @@
 import post from './api/posts';
 import user from './api/users';
 import comment from './api/comments';
+import authentication from './authentication';
 
 export default {
   post,
   user,
-  comment
+  comment,
+  authentication
 };
