+++ apps/personal-blog-backend/src/main.ts
@@ -6,8 +6,6 @@
 import * as express from 'express';
 import * as cors from 'cors';
 import * as bodyParser from 'body-parser';
-import * as jwt from 'jsonwebtoken';
-import * as bcrypt from 'bcrypt';
 
 import models, { connectDb } from './app/models';
 
@@ -28,27 +26,7 @@ app.use(cors());
 app.use('/api/posts', routes.post);
 app.use('/api/users', routes.user);
 app.use('/api/comments', routes.comment);
-
-app.post('/api/authenticate', async (req, rest, next) => {
-  const body = req.body;
-  models.User.findOne({ email: body.email }, (err, user) => {
-    if (err) {
-      next(err);
-    } else {
-      bcrypt.compare(body.password, user.password, function(passErr, isMatch) {
-        if (passErr) return next(passErr);
-        const token = jwt.sign({ email: user.email }, 'shhhhh');
-        rest.send({
-          _id: user._id,
-          email: user.email,
-          auth_token: token,
-          avatar: user.avatar,
-          name: user.name || user.email
-        });
-      });
-    }
-  });
-});
+app.use('/api/authenticate', routes.authentication);
 
 const port = process.env.port || 3333;
